# Improve performance of lox-rs by avoiding clone operations