--- a/lox-programs/fib-rec.lox
+++ b/lox-programs/fib-rec.lox
@@ -4,7 +4,7 @@ fun fib(n) {
   return fib(n - 2) + fib(n - 1);
 }
 
-for (var i = 0; i < 3; i = i + 1) {
+for (var i = 0; i < 28; i = i + 1) {
   print fib(i);
   print "returned";
 }