--- a/lox-programs/fib.lox
+++ b/lox-programs/fib.lox
@@ -1,8 +1,10 @@
 var a = 0;
 var temp;
 
-for (var b = 1; a < 10000; b = temp + b) {
-  print a;
-  temp = a;
-  a = b;
+for (var c = 1; c < 1000000; c = c + 1) {
+    for (var b = 1; a < 10000; b = temp + b) {
+      print a;
+      temp = a;
+      a = b;
+    }
 }