@ntietz 2022-11-23
Hopefully the build passes after the latest commits. I had to update `clippy` and there were some new lints.